攝氏轉換華氏

+++ cf.py
@@ -1,4 +1,4 @@
 C = input('請輸入攝氏')
-C = int(C)
+C = float(C)
 F = 9 / 5 * C + 32
 print (F)
